fix waste pr quality gates
