+++ packages/plugin-waste-management/src/waste-management.master-data-fraction-dialog.fields.tsx
@@ -1,0 +1,165 @@
+import { usePluginTranslation } from '@sva/plugin-sdk';
+import {
+  Input,
+  StudioField,
+  StudioFieldGroup,
+  StudioFormSummaryErrors,
+  getStudioFormFieldProps,
+} from '@sva/studio-ui-react';
+import { Controller, useForm } from 'react-hook-form';
+
+import { WasteManagementFormSwitch } from './waste-management.form-switch.js';
+import type { FractionFormState } from './waste-management.master-data.forms.js';
+import { collectSummaryErrors, type BaseProps } from './waste-management.master-data-entity-dialogs.shared.js';
+
+const FractionDialogTextFields = ({
+  clearErrors,
+  nameField,
+  onChange,
+  pdfShortLabel,
+  pdfShortLabelField,
+  pt,
+  register,
+  translations,
+}: {
+  readonly clearErrors: ReturnType<typeof useForm<FractionFormState>>['clearErrors'];
+  readonly nameField: ReturnType<typeof getStudioFormFieldProps>;
+  readonly onChange: BaseProps<FractionFormState>['onChange'];
+  readonly pdfShortLabel: FractionFormState['pdfShortLabel'];
+  readonly pdfShortLabelField: ReturnType<typeof getStudioFormFieldProps>;
+  readonly pt: ReturnType<typeof usePluginTranslation>;
+  readonly register: ReturnType<typeof useForm<FractionFormState>>['register'];
+  readonly translations: FractionFormState['translations'];
+}) => (
+  <StudioFieldGroup>
+    <StudioField {...nameField} label={pt('masterData.fractions.fields.name')}>
+      <Input
+        {...nameField.controlProps}
+        {...register('name', {
+          required: pt('masterData.fractions.fields.name'),
+          onChange: (event) => {
+            clearErrors('name');
+            onChange({ name: event.target.value });
+          },
+        })}
+      />
+    </StudioField>
+    <StudioField {...pdfShortLabelField} label={pt('masterData.fractions.fields.pdfShortLabel')}>
+      <Input
+        {...pdfShortLabelField.controlProps}
+        value={pdfShortLabel}
+        maxLength={12}
+        {...register('pdfShortLabel', {
+          required: pt('masterData.fractions.fields.pdfShortLabel'),
+          onChange: (event) => {
+            clearErrors('pdfShortLabel');
+            onChange({ pdfShortLabel: event.target.value });
+          },
+        })}
+      />
+    </StudioField>
+    <StudioField id="waste-fraction-name-de" label={pt('masterData.fractions.fields.translationDe')}>
+      <Input
+        id="waste-fraction-name-de"
+        {...register('translations.de' as const, {
+          onChange: (event) => onChange({ translations: { ...translations, de: event.target.value } }),
+        })}
+      />
+    </StudioField>
+    <StudioField id="waste-fraction-name-en" label={pt('masterData.fractions.fields.translationEn')}>
+      <Input
+        id="waste-fraction-name-en"
+        {...register('translations.en' as const, {
+          onChange: (event) => onChange({ translations: { ...translations, en: event.target.value } }),
+        })}
+      />
+    </StudioField>
+    <StudioField id="waste-fraction-color" label={pt('masterData.fractions.fields.color')}>
+      <Input id="waste-fraction-color" {...register('color', { onChange: (event) => onChange({ color: event.target.value }) })} />
+    </StudioField>
+    <StudioField id="waste-fraction-container-size" label={pt('masterData.fractions.fields.containerSize')}>
+      <Input id="waste-fraction-container-size" {...register('containerSize', { onChange: (event) => onChange({ containerSize: event.target.value }) })} />
+    </StudioField>
+    <StudioField id="waste-fraction-description" label={pt('masterData.fractions.fields.description')}>
+      <Input id="waste-fraction-description" {...register('description', { onChange: (event) => onChange({ description: event.target.value }) })} />
+    </StudioField>
+  </StudioFieldGroup>
+);
+
+const FractionDialogActiveField = ({
+  active,
+  control,
+  onChange,
+  pt,
+}: {
+  readonly active: boolean;
+  readonly control: ReturnType<typeof useForm<FractionFormState>>['control'];
+  readonly onChange: BaseProps<FractionFormState>['onChange'];
+  readonly pt: ReturnType<typeof usePluginTranslation>;
+}) => (
+  <div className="flex items-center gap-3">
+    <Controller
+      name="active"
+      control={control}
+      render={({ field }) => (
+        <WasteManagementFormSwitch
+          checked={field.value}
+          ariaLabel={pt('masterData.fractions.fields.active')}
+          onChange={(nextActive) => {
+            field.onChange(nextActive);
+            onChange({ active: nextActive });
+          }}
+        />
+      )}
+    />
+    <span className="text-sm text-muted-foreground">{active ? pt('common.active') : pt('common.inactive')}</span>
+  </div>
+);
+
+export const FractionDialogFields = ({
+  active,
+  clearErrors,
+  control,
+  errors,
+  onChange,
+  pdfShortLabel,
+  pt,
+  register,
+  translations,
+}: {
+  readonly active: boolean;
+  readonly clearErrors: ReturnType<typeof useForm<FractionFormState>>['clearErrors'];
+  readonly control: ReturnType<typeof useForm<FractionFormState>>['control'];
+  readonly errors: ReturnType<typeof useForm<FractionFormState>>['formState']['errors'];
+  readonly onChange: BaseProps<FractionFormState>['onChange'];
+  readonly pdfShortLabel: FractionFormState['pdfShortLabel'];
+  readonly pt: ReturnType<typeof usePluginTranslation>;
+  readonly register: ReturnType<typeof useForm<FractionFormState>>['register'];
+  readonly translations: FractionFormState['translations'];
+}) => {
+  const nameField = getStudioFormFieldProps({
+    id: 'waste-fraction-name',
+    error: errors.name,
+  });
+  const pdfShortLabelField = getStudioFormFieldProps({
+    id: 'waste-fraction-pdf-short-label',
+    error: errors.pdfShortLabel,
+  });
+
+  return (
+    <>
+      <StudioFormSummaryErrors errors={collectSummaryErrors([nameField, pdfShortLabelField])} />
+      <FractionDialogTextFields
+        clearErrors={clearErrors}
+        nameField={nameField}
+        onChange={onChange}
+        pdfShortLabel={pdfShortLabel}
+        pdfShortLabelField={pdfShortLabelField}
+        pt={pt}
+        register={register}
+        translations={translations}
+      />
+      <FractionDialogActiveField active={active} control={control} onChange={onChange} pt={pt} />
+    </>
+  );
+};
